+++ deviceapp/src/ledController.h
@@ -5,23 +5,23 @@
 
 class LedController {
 public:
-    LedController(int pin, int numLedsX, int numLedsY);
+    LedController(uint8_t pin, uint8_t numLedsX, uint8_t numLedsY);
     void setup();
-    void setColor(int r, int g, int b);
-    void setColor(int r, int g, int b, int ledX, int ledY);
-    void setBrightness(int brightness);
+    void setColor(uint8_t r, uint8_t g, uint8_t b);
+    void setColor(uint8_t r, uint8_t g, uint8_t b, uint8_t ledX, uint8_t ledY);
+    void setBrightness(uint8_t brightness);
     void clear();
-    void clear(int ledX, int ledY);
+    void clear(uint8_t ledX, uint8_t ledY);
     void show();
     void noConnection();
     void yesHotSpotNoWSConnection(String ip);
     void allConnection();
 
-    int assConverter(int index);
+    uint16_t assConverter(uint16_t index);
 
 private:
     Adafruit_NeoPixel strip;
-    int width, height; // Dimensions of the matrix
+    uint8_t width, height; // Dimensions of the matrix
 };
 
 #endif
